requirements.txt
Flask==2.2.5
pandas==2.0.0
scikit-learn==1.4.2
plotly==5.3.1
numpy==1.24.4
gunicorn==23.0.0
psycopg2-binary==2.9.1
python-dotenv==0.19.2
SQLAlchemy==2.0.16
Werkzeug==2.2.2
